Added games played tracking

+++ events/interactionCreate.js
@@ -2,6 +2,7 @@ const { LineupQueue, Challenge } = require("../mongoSchema");
 const interactionUtils = require("../services/interactionUtils");
 const matchmakingService = require("../services/matchmakingService");
 const teamService = require("../services/teamService");
+const statsService = require("../services/statsService");
 const authorizationService = require("../services/authorizationService");
 
 module.exports = {
@@ -57,18 +58,16 @@ module.exports = {
                     lineup = await teamService.findLineupByChannelId(interaction.guildId, interaction.channelId)
                     let numberOfPlayersSigned = lineup.roles.filter(role => role.user != null).length
                     let missingRoleName = lineup.roles.find(role => role.user == null)?.name
-                    if (lineup.autoSearch === true && numberOfPlayersSigned == lineup.roles.length || (numberOfPlayersSigned >= lineup.roles.length-1 && missingRoleName === 'GK')) {
+                    if (lineup.autoSearch === true && (numberOfPlayersSigned == lineup.roles.length || (numberOfPlayersSigned >= lineup.roles.length - 1 && missingRoleName === 'GK'))) {
                         let lineupQueue = await matchmakingService.findLineupQueueByChannelId(interaction.channelId)
-                        if (lineupQueue) {
-                            await interactionUtils.replyAlreadyQueued(interaction, lineupQueue.lineup.size)
+                        if (!lineupQueue) {
+                            await new LineupQueue({
+                                team: team,
+                                lineup: lineup
+                            }).save()
+                            await interaction.reply(`Player ${interaction.user} signed into the lineup as ${roleName}. Your lineup is full, it is now queued for ${lineup.size}v${lineup.size} !`)
                             return
                         }
-                        await new LineupQueue({
-                            team: team,
-                            lineup: lineup
-                        }).save()
-                        await interaction.reply(`Player ${interaction.user} signed into the lineup as ${roleName}. Your lineup is full, it is now queued for ${lineup.size}v${lineup.size} !`)
-                        return
                     }
 
                     await interaction.reply({ content: `Player ${interaction.user} signed into the lineup as ${roleName}`, components: interactionUtils.createLineupComponents(lineup, interaction.user.id) })
@@ -79,7 +78,7 @@ module.exports = {
                     let existingPlayerRole = lineup.roles.find(role => role.user?.id === interaction.user.id)
 
                     if (!existingPlayerRole) {
-                        await interaction.reply({ content: `❌ You are not in the lineup` , ephemeral: true})
+                        await interaction.reply({ content: `❌ You are not in the lineup`, ephemeral: true })
                         return
                     }
 
@@ -91,9 +90,12 @@ module.exports = {
                     lineup = await teamService.findLineupByChannelId(interaction.guildId, interaction.channelId)
 
                     if (lineup.autoSearch === true && lineup.roles.filter(role => role.user != null).length <= lineup.roles.length - 1) {
-                        await LineupQueue.deleteOne({ 'lineup.channelId': interaction.channelId })
-                        await interaction.reply({ content: `Player ${interaction.user} left the ${existingPlayerRole.name} position. Your team is no longer in the queue !`, components: interactionUtils.createLineupComponents(lineup, interaction.user.id) })
-                        return
+                        let challenge = await matchmakingService.findChallengeByGuildId(interaction.guildId)
+                        if (!challenge) {
+                            await LineupQueue.deleteOne({ 'lineup.channelId': interaction.channelId })
+                            await interaction.reply({ content: `Player ${interaction.user} left the ${existingPlayerRole.name} position. Your team is no longer in the queue !`, components: interactionUtils.createLineupComponents(lineup, interaction.user.id) })
+                            return
+                        }
                     }
 
                     await interaction.reply({ content: `Player ${interaction.user} left the ${existingPlayerRole.name} position`, components: interactionUtils.createLineupComponents(lineup, interaction.user.id) })
@@ -173,6 +175,9 @@ module.exports = {
 
                     await interaction.message.edit({ components: [] })
                     await interaction.reply(`⚽ You have accepted to challenge the team '${teamService.formatTeamName(challenge.initiatingTeam.team, challenge.initiatingTeam.lineup)}' ! Check your private messages for lobby info !`)
+
+                    await statsService.incrementGamesPlayed(users)
+
                     return
                 }
